Move acorn to javascript module
