+++ python/jsbeautifier/core/tokenizer.py
@@ -43,19 +43,14 @@ TOKEN = TokenTypes()
 class Tokenizer:
 
     def __init__(self, input_string, options):
-        import jsbeautifier.core.acorn as acorn
-        self.acorn = acorn
-
         self._input = InputScanner(input_string)
         self._options = options
         self.__tokens = None
         self.__newline_count = 0
         self.__whitespace_before_token = ''
 
-        self._whitespace_pattern = re.compile(
-            self.acorn.six.u(r'[\n\r\u2028\u2029\t\u000B\u00A0\u1680\u180e\u2000-\u200a\u202f\u205f\u3000\ufeff ]+'))
-        self._newline_pattern = re.compile(
-            self.acorn.six.u(r'([^\n\r\u2028\u2029]*)(\r\n|[\n\r\u2028\u2029])?'))
+        self._whitespace_pattern = re.compile(r'[\n\r\t ]+')
+        self._newline_pattern = re.compile(r'([^\n\r]*)(\r\n|[\n\r])?')
 
     def tokenize(self):
         self._input.restart()
